initial godoctor plugin support for extract func
this adds support for https://github.com/godoctor/godoctor in a relatively
general manner in the backend, so that we might extend and add more
functionality relatively easily in the future. initially, wrt vim-go, only
:GoExtractFunc has been added for the time being, and :GoInstallBinaries and
:GoUpdateBinaries will also be integrated to download/update the godoctor.

we have support for a vim plugin at https://github.com/godoctor/godoctor.vim
of our own but the Go community seems to have settled on the wonderful vim-go
package, so we would prefer to integrate into vim-go, if possible (if so, will
port most of this over to that and we will mirror our vim-go version to any
users who may just want standalone) -- this made it relatively easy to move
over. the godoctor itself has support for a few other refactorings already,
but wanted to test the waters here before going whole hog. it would also be
interesting to discuss how we might like to add support for our rename
functionality as well as the gorename functionality (I have found a few cases
over time where one works and the other doesn't -- both ways!). I was thinking
along the lines of :GoDoctorRename or such, if this is desirable at all (I can
see the confusion..).

tried my best to port most of our helper funcs to use vim-go's tooling and
output, but it may very well be that I have come up a little short (style,
etc). please let me know if there is any further tidying up to be done as I
was not able to find a contributor guide I simply tried my best! I tested
sync/async with errors and not and initially looks pretty good, there's a
warning from the godoctor about scoping on success but I'm going to take care
of that in the godoctor itself, it does succeed and should be pretty obvious
for extract func that it has; errors populate the quickfix list and look good
(even though they are bad!).

testing should be as easy as selecting a line (or a few) in visual mode and then doing:

```
:'<,'>GoExtractFunc yodawg
```

(the '<,'> will (should?) get auto populated after typing `:`)

thank you for vim-go and all your awesome work on it, we look forward to
integrating with it!